asca

+++ segmentationmodule.py
@@ -108,10 +108,10 @@ class SegmentCyst(pl.LightningModule):
 
 
             ax1.set_title('IMAGE')
-            ax1.imshow(image.cpu().permute(1,2,0).squeeze(),cmap = 'gray')
+            ax1.imshow(image.cpu().permute(1,2,0).squeeze(),cmap = 'gray').astype(np.uint8)
 
             ax2.set_title('GROUND TRUTH')
-            ax2.imshow((y_true * 255).cpu().permute(1,2,0).squeeze(),cmap = 'gray')
+            ax2.imshow((y_true * 255).cpu().permute(1,2,0).squeeze(),cmap = 'gray').astype(np.uint8)
 
             """ ax3.set_title('MODEL OUTPUT')
             y_pred = (y_pred > 0.5).permute(1,2,0).cpu().detach().numpy().astype(np.uint8)
